requirements.txt
aiosqlite==0.17.0
fastapi==0.95.1
openai==0.27.5
python-dotenv==1.0.0
tortoise-orm==0.19.3
uvicorn==0.22.0
